Add option to cache ASTs
When parsing repeatedly and changing wiki articles that are rather late in
the parse cycle, this can speed up the speed to the first error from about
30 seconds (on my machine) to less than 5 seconds, allowing shorter wait
times and a faster turnaround.

Hid this behind an option as it slows down the initial parsing and you
have to think about deleting the cache when you change the parser.