Add option to use previous versions
When parsing fails, and the new -p option is given, try to retrieve the
previous version of the wiki page and parse it instead. This continues
until either there are no more revision, or the revision is older than
minHistoryDate (currently 2015-01-01). In that case, an empty page (just
as if the article did not exist yet) is used.

That way, it should be (unless the parser is changed in incompatible ways)
always possible to create a (best-effort) export file containing as many
and as recent chapters as possible.

The retrieved previous page overwrites the cached page in the page cache;
therefore the history walk will not happen again when re-running without
purging the page cache.