Make OSIS valid according to schema
I'm pretty sure that is not the way how it is intended, but given that <l>
can only appear inside <lg> and that <lg> may not contain <note> (or,
according to the errata, mixed content at all), I can't think of a better
way to fix.

Note that this has NOT been tested against SWORD; no idea how they will
handle it.