Convert Zefania XML to e-Sword Tool HTML files
e-Sword uses an authoring tool called "ToolTip Tool NT" that can import
specially formatted HTML and RTF files and convert them to bblx and cmtx
files.

This converter converts the Zefania XML files for Offene Bibel to this
HTML format. Note that e-Sword does not support footnotes in bible files;
therefore every verse that has footnotes also needs a commentary entry in
the commentary file (which will be shown when the verse is clicked) that
contains the footnotes.

That way, e-Sword modules for Offene Bibel can be created.

Expect some very long conversion times of the "ToolTip Tool" (while this
script runs in a fraction of a second, the actual conversion takes several
minutes); also, the import feature sometimes inserts newlines in the
middle of words, causing a parse error when exporting as BBLX. To fix,
either try to import the HTML again (in the hope that it will work, as the
HTML import seems to be non-deterministic) or manually fix the newline in
the authoring tool and export again (until all errors are fixed and no
error is shown).

As some verse of Offene Bibel are outside the Canon used by e-Sword,
these verses are stripped from the export since otherwise they would all
have to be manually deleted in the authoring tool.