Tag addition, omission and alternative in OSIS
Use `<transChange>` for addition and omission (as specified), but use
`<seg type="x-alternative">` for alternatives, as there is no specified
way to tag it, and as transChanges may not be nested but alternatives are
often nested inside or around other transChanges.

Also ensure that transChanges within other transChanges (currently 3
occurrences in the whole Offene Bibel) will not be tagged, to avoid a
validation error. Do not add seg or transChange within notes, either.

transChange and seg are currently ignored for Zefania XML.